create video server module
